+++ src/interfaces/plugins/index.ts
@@ -2,6 +2,7 @@ export * from './interactions/interactions.interface';
 export * from './keyboardHandler/keyboardHandler.interface';
 export * from './liveSearch/liveSearch.interface';
 export * from './normalState/normalState.interface';
+export * from './normalState/simpleNormalState.interface';
 export * from './optionsHandler/optionsHandler.interface';
 export * from './popup/popup.interface';
 export * from './positioner/positioner.interface';