+++ src/interfaces/plugins/readonlyState/readonlyState.interface.ts
@@ -1,9 +1,16 @@
-import {NormalState, NormalStateOptions} from '../normalState/normalState.interface';
+import {NormalState, NormalStateCssClasses, NormalStateOptions} from '../normalState/normalState.interface';
+
+/**
+ * CSS classes for readonly state plugin
+ */
+export interface ReadonlyStateCssClasses extends NormalStateCssClasses
+{
+}
 
 /**
  * Options for readonly state plugin
  */
-export interface ReadonlyStateOptions<TCssClasses = unknown> extends NormalStateOptions<TCssClasses>
+export interface ReadonlyStateOptions<TCssClasses extends ReadonlyStateCssClasses = ReadonlyStateCssClasses> extends NormalStateOptions<TCssClasses>
 {
 }