+++ src/pipes/index.ts
@@ -1,3 +1,3 @@
+export * from './displayValue/displayValue.pipe';
 export * from './getPlugin/getPlugin.pipe';
 export * from './selectHasValue/selectHasValue.pipe';
-export * from './selectValue/selectValue.pipe';