+++ src/plugins/normalState/simpleNormalState/simpleNormalState.component.ts
@@ -1,29 +1,73 @@
-import {Component, ElementRef} from '@angular/core';
+import {ChangeDetectionStrategy, Component, ElementRef, Inject, inject, Optional} from '@angular/core';
+import {NgTemplateOutlet} from '@angular/common';
+import {LocalizePipe, TooltipDirective} from '@anglr/common';
+import {RecursivePartial} from '@jscrpt/common';
+import {deepCopyWithArrayOverride} from '@jscrpt/common/lodash';
 
-import {NormalState} from '../../../interfaces';
+import {NormalState, NormalStateOptions, SelectPlugin, SimpleNormalStateCssClasses} from '../../../interfaces';
 import {SelectPluginInstances, SelectBus} from '../../../misc/classes';
+import {CopyOptionsAsSignal} from '../../../decorators';
+import {NORMAL_STATE_OPTIONS} from '../../../misc/tokens';
+import {DisplayValue} from '../../../pipes';
+
+//TODO: improvement, carret direction
+
+const defaultOptions: NormalStateOptions<SimpleNormalStateCssClasses> =
+{
+    cssClasses:
+    {
+        containerElement: '',
+        value: '',
+        carret: 'fas fa-caret-down',
+    },
+};
 
 /**
- *
+ * Simple normal state displaying value as 'button' that opens popup with options when clicked.
  */
 @Component(
 {
     selector: 'simple-normal-state',
-    template: '',
+    templateUrl: 'simpleNormalState.component.html',
+    imports:
+    [
+        DisplayValue,
+        LocalizePipe,
+        TooltipDirective,
+        NgTemplateOutlet,
+    ],
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class SimpleNormalStateComponent implements NormalState
+export class SimpleNormalState<TValue = unknown> implements NormalState<TValue, NormalStateOptions<SimpleNormalStateCssClasses>>
 {
+    //######################### public properties - implementation of SelectPlugin #########################
+
     /**
      * @inheritdoc
      */
-    public options: unknown;
+    @CopyOptionsAsSignal()
+    public options: NormalStateOptions<SimpleNormalStateCssClasses>;
+
+    /**
+     * @inheritdoc
+     */
+    public selectPlugins: SelectPluginInstances = inject(SelectPluginInstances);
+
+    /**
+     * @inheritdoc
+     */
+    public pluginElement: ElementRef<HTMLElement> = inject(ElementRef);
+
+    /**
+     * @inheritdoc
+     */
+    public selectBus: SelectBus<TValue> = inject(SelectBus);
 
     //######################### constructor #########################
-    constructor(public pluginElement: ElementRef<HTMLElement>,
-                public selectPlugins: SelectPluginInstances,
-                public selectBus: SelectBus<unknown>,)
+    constructor(@Inject(NORMAL_STATE_OPTIONS) @Optional() options?: RecursivePartial<NormalStateOptions<SimpleNormalStateCssClasses>>|null,)
     {
-
+        this.options = deepCopyWithArrayOverride(defaultOptions as NormalStateOptions<SimpleNormalStateCssClasses>,
+                                                 options);
     }
 
     //######################### public methods - implementation of SelectPlugin #########################
@@ -40,5 +84,31 @@ export class SimpleNormalStateComponent implements NormalState
      */
     public initOptions(): void
     {
+    }
+
+    //######################### protected methods - template bindings #########################
+
+    /**
+     * Handles click event
+     */
+    protected click(): void
+    {
+        this.selectBus.click.next(
+        {
+            source: this as SelectPlugin,
+            sourceElement: this.pluginElement.nativeElement,
+        });
+    }
+
+    /**
+     * Handles focus event
+     */
+    protected focus(): void
+    {
+        this.selectBus.focus.next(
+        {
+            source: this as SelectPlugin,
+            sourceElement: this.pluginElement.nativeElement,
+        });
     }
 }